ModelView Refactoring
