+++ Fitnessapp/Models/Activity.swift
@@ -1,0 +1,17 @@
+//
+//  Activity.swift
+//  Fitnessapp
+//
+//  Created by Sieg Friedrich on 2024/12/09.
+//
+
+import SwiftUI
+
+struct Activity {
+    let id: Int
+    let title: String
+    let subtitle: String
+    let image: String
+    let tintColor: Color
+    let amount: String
+}
