+++ Fitnessapp/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  Fitnessapp
+//
+//  Created by Sieg Friedrich on 2024/12/09.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id: Int
+    let image: String
+    let tintColor: Color
+    let title: String
+    let duration: String
+    let date: String
+    let calories: String
+}
