+++ Fitnessapp/Views/ActivityCard.swift
@@ -6,14 +6,7 @@
 //
 
 import SwiftUI
-struct Activity {
-    let id: Int
-    let title: String
-    let subtitle: String
-    let image: String
-    let tintColor: Color
-    let amount: String
-}
+
 
 struct ActivityCard: View {
