+++ Fitnessapp/Views/HistoricDataView.swift
